requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==4.0.0
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
entrypoints==0.4
executing==2.0.1
faiss-cpu==1.7.4
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.0
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
langchain==0.0.316
langchain-community==0.0.17
langchain-core==0.1.18
langsmith==0.0.85
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.3
openai==0.28.1
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.4
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
semver==3.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
st-annotated-text==4.0.1
stack-data==0.6.3
streamlit==1.18.1
streamlit-camera-input-live==0.2.0
streamlit-embedcode==0.1.2
streamlit-extras==0.1.5
streamlit-keyup==0.2.2
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.13
yarl==1.9.4
zipp==3.17.0
